+++ jgltf-model/src/main/java/de/javagl/jgltf/model/v1/GltfModelCreatorV1.java
@@ -473,7 +473,7 @@ public class GltfModelCreatorV1
         for (Entry<String, Skin> entry : skins.entrySet())
         {
             Skin skin = entry.getValue();
-            float[] bindShapeMatrix = skin.getBindShapeMatrix();
+            double[] bindShapeMatrix = skin.getBindShapeMatrix();
             DefaultSkinModel skinModel = new DefaultSkinModel();
             skinModel.setBindShapeMatrix(bindShapeMatrix);
             gltfModel.addSkinModel(skinModel);
@@ -1007,11 +1007,11 @@ public class GltfModelCreatorV1
                     get("cameras", cameraId, gltfModel::getCameraModel);
                 nodeModel.setCameraModel(cameraModel);
             }
-            
-            float matrix[] = node.getMatrix();
-            float translation[] = node.getTranslation();
-            float rotation[] = node.getRotation();
-            float scale[] = node.getScale();
+
+            double matrix[] = node.getMatrix();
+            double translation[] = node.getTranslation();
+            double rotation[] = node.getRotation();
+            double scale[] = node.getScale();
             nodeModel.setMatrix(Optionals.clone(matrix));
             nodeModel.setTranslation(Optionals.clone(translation));
             nodeModel.setRotation(Optionals.clone(rotation));
